Adds IAM module for GitHub Actions deployments
Introduces a new module to create an IAM user specifically for GitHub Actions deployments.

This module facilitates automated deployments by providing the necessary permissions to interact with S3 and CloudFront. The IAM user will be granted access to upload files to the specified S3 bucket and invalidate the CloudFront cache, ensuring smooth and automated deployments.